publish pipeline: restore NovelSourcery branding and isNovel emission
kei folded create-repo.py + merge-repo.py into a single publish-repo.py and
switched its metadata source from parsing `aapt dump badging` output to
reading the JSON GenerateSourceInfoTask now emits at build time. The new
script was a clean copy of kei's own upstream version -- Keiyoushi branding,
their signing key, their CDN URLs, generic `tachiyomi-` naming, and (this is
the one that actually mattered) never sets isNovel on the extensions it
publishes, since the manifest-meta-data-tag mechanism that used to derive it
doesn't exist in the new pipeline at all.

- index.proto: added the missing `isNovel = 8000` field declaration (the fork's
  reserved 8000+ field block) -- the compiled index_pb2.py/.pyi already had it
  from an earlier session, hand-patched without updating the .proto source they're
  supposed to be generated from, so the two had drifted apart. Regenerated both
  via `protoc --python_out=. --pyi_out=.` per the file's own header instructions,
  confirmed byte-identical in the only intentional aspect (offsets) and verified
  with a live strict-JSON parse round-trip.
- publish-repo.py: name/badgeLabel/signingKey/contact -> NovelSourcery's (from
  the old merge-repo.py), LANGUAGE_REGEX and the apk/jar/icon cleanup globs
  tachiyomi- -> tsundoku- and eu.kanade.tachiyomi.extension. ->
  eu.kanade.tachiyomi.novelextension. (matching the fork's actual APK naming
  convention), and `isNovel=info.get("isNovel", True)` threaded through to
  index_pb2.Extension -- paired with the ExtensionPlugin.kt commit's
  `isNovel: true` on the JSON payload this reads.
  APK/JAR/ICON_BASE_URL kept on the fork's existing raw.githubusercontent.com
  pattern rather than adopting kei's new jsdelivr CDN switch -- a deliberate,
  conservative call (only found evidence jsdelivr is intentionally used for
  index.min.json specifically, via commit-repo.sh's purge call, not for
  APK/icon downloads), flagging in case that's wrong and jsdelivr should
  apply here too.

Deleted create-repo.py, merge-repo.py, commit-repo.sh, and always_build.json
now that publish-repo.py fully replaces them and nothing else references them.
