--- .github/always_build.json
@@ -1,2 +1,0 @@
-[
-]
