+++ .github/scripts/generate-build-matrices.py
@@ -5,7 +5,6 @@ import re
 import subprocess
 import sys
 from pathlib import Path
-from typing import NoReturn
 
 EXTENSION_REGEX = re.compile(r"^src/(?P<lang>\w+)/(?P<extension>\w+)")
 MULTISRC_LIB_REGEX = re.compile(r"^lib-multisrc/(?P<multisrc>\w+)")
@@ -183,13 +182,6 @@ def get_module_list(ref: str) -> tuple[list[str], list[str]]:
     extensions = resolve_ext(multisrcs, libs)
     modules.update([f":src:{lang}:{extension}" for lang, extension in extensions])
     deleted.update([f"{lang}.{extension}" for lang, extension in extensions])
-
-    if os.getenv("IS_PR_CHECK") != "true":
-        with Path(".github/always_build.json").open() as always_build_file:
-            always_build = json.load(always_build_file)
-        for extension in always_build:
-            modules.add(":src:" + extension.replace(".", ":"))
-            deleted.add(extension)
 
     return list(modules), list(deleted)
